fix: rename image in assets more descriptive.
